requirements.txt
monai
mrcfile
pyyaml
matplotlib
numpy
